+++ lib/presentation/view/publisher_view.dart
@@ -36,10 +36,10 @@ class _PublisherViewState extends State<PublisherView> {
             FadeInImage.assetNetwork(
               placeholder: "assets/images/placeholder.png",
               image: widget.result.imageBackground,
-              imageCacheHeight: _config.sh(450),
-              imageCacheWidth: _config.sw(800),
-              placeholderCacheHeight: _config.sh(400),
-              placeholderCacheWidth: _config.sw(400),
+              imageCacheHeight: _config.sh(450).toInt(),
+              imageCacheWidth: _config.sw(800).toInt(),
+              placeholderCacheHeight: _config.sh(400).toInt(),
+              placeholderCacheWidth: _config.sw(400).toInt(),
               fit: BoxFit.cover, height: _config.sh(200),
               width: SizeConfig.screenWidthDp,),
             Container(
